+++ py7zr/tests/test_basic.py
@@ -1,3 +1,4 @@
+import binascii
 import io
 import os
 import lzma
@@ -272,7 +273,8 @@ def test_lzma_lzma2bcj_compressor():
 @pytest.mark.unit
 def test_read_archive_properties():
     buf = io.BytesIO()
-    buf.write(b'02070123456789abcd00')
+    inp = binascii.unhexlify('0207012300')
+    buf.write(inp)
     buf.seek(0, 0)
     ap =  py7zr.archiveinfo.ArchiveProperties.retrieve(buf)
-    assert ap.property_data[0] == b'0123456789abcd'
+    assert ap.property_data[0] == (0x23, )  # FIXME: what should be?